Build Markovs On Startup
For discord guilds with large text databases, the bot took too long to
respond. This is a result of the fact that we were parsing and building
the markov chain every time the command was issued. Now we build the
structure on startup, and reuse each `.Walk()` until we're at the end,
and then we simply rebuild for a new `.Walk()`